requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2
channels
channels-redis
redis
django-redis
cryptography
pycryptodome
daphne
boto3